Stop URL decoding the URL when fetching a cache copy.
This doesn't make sense and causes problems when the URL contains encoded chars (e.g. we convert `http://foo/a%20b/` to `http://foo/a b/` and then find we can't find that URL in the warc file as it is in the correct
t `http://foo/a%20b/` form.