Mark faceted nav extra searches so that hook scripts know when they are dealing with them
FUN-11166

This adds support for hooks scripts to detect if they are running on a faceted nav extra search. This way they have the option to opt out of manipulating the search transaction if it is going to cause issues for example setting num_ranks=10000 on all of the extra searches would not be great.