Add support for preventing a query from being logged
FUN-11275 SUPPORT-2614

It is now possible to set on the SearchQuestion if the query should be logged or not.
```
// e.g. to turn off logging
transaction.question.setLogQuery(false);
```

When turned off queries will not be logged in the query log or in search session history.

This option can be set via CGI params as long as it is done in a authenticated request. To do this set `&log=off`.

I have updated a stack of places which set the command line option to padre `-log=off` to now set it on the search question.

Faceted navigation extra searches now has it set to false by default.