Allow setting -log in collection.cfg and other files
FUN-11275 SUPPORT-2614

An issue existed where logging from padre-sw could not be controlled from collection.cfg because we always set the value when calling padre.
This changes `logQuery` to be an `Optional` when empty the behavior of logging is left up to padre as it was before, which means it will respect config files.

Also include `FacetedNavigationQuestionFactoryTest` missing from the last MR.