Fix Search history signature generation from double decoding the URL query
FUN-11396 SUPPORT-2630

When the query contains a `%` which is represented in the url as
```
query=%25
```
this will be first decoded to `%` by
`com.funnelback.publicui.utils.QueryStringUtils.toArrayMap(String)`
and was being decoded again in
`com.funnelback.publicui.utils.URLSignature.computeQueryStringSignature(String)`
. This could sometimes cause issue.

The change is to remove the second decode. Note that this has a minor
issue where old search sessions may not match up because the query hash
has changed. I think this should be ok as search sessions can just be
cleared.