Initial support for fetching related items into the search data model.
Relations are represented as metadata on a document where the value is
a list of URLs of the related documents in the index, a form which
padre-i4u should be able to efficiently handle.

Lots left to do here, but I think it shows the basic concept.

(Motivating example is slack search, where we want to store each
message as its own document, but have a way to present the priorMessage
if there is one in the search results for context).