Some fixes to help with maintaining the casing of smb paths
Move the URL fix logic our to `FacetURL`. This then gives access to two
forms of the URL one which is fixed e.g. `\\Foo\Bar` to `smb://Foo/Bar/` and
another which is used for comparison where case insensitive parts are
lowercased e.g. `smb://foo/bar/`. This makes it easy to do comparisons
while also making it possible to preserve case when presenting values to
users.