Reduce the memory used by the history purge API call.
I tested this for a while by changing the amount of memory used in the
data created for the SearchHistoryDaoPurgeTest (adding long
summaries, metadata and lots of records). It seems like the
regular flush calls make a big difference, and locgically the
ScrollableResults should help though the old version seems not
to behave the way I read the code, so maybe it's not as crazy as
it looks.

See: SUPPORT-2724