Simplifies the Cart result json
now it is:
```
{

    "userId": "769b2023-5fab-414e-b5d9-3ca429091e5a",
    "collection": "Se2-PublicUI-Session",
    "indexUrl": "http://test-data.cbr.au.funnelback.com/Shakespeare/lear/lear.1.3.html",
    "title": "SCENE III. The Duke of Albany's palace.",
    "summary": "King Lear...",
    "addedDate": 1532915317754,
    "metaData": {
        "FunResponseTime": "92",
        "FunResponseTimeBand": "21-100"
    }

},
```

The collection will typically be set to the collection which has the indexUrl.

In some cases when working with old data collection may still be the meta collection but after a few weeks all of the session data should be re-freshed and it will go to the component collection.

The DB stores a diferent model to what is presented to the user hence the creation of `CartResultDBModel`.