Fixes a memory leak
The final list is mutable and so was being muated to have a new value of the daat limit on every query. Eventually this would lead to a OOM, the fix here is to just remove the final field and generate all the options on the fly. The field has no comment about it needed to be pre-computed so I think it is safe to do this.