Added a Angular filter to replace strings in template
When using the Angular shortlist templates it's very frequent that we
need to replace `|` with `, ` with multi-valued metadata (as the default
separator is a pipe).

The Angular template syntax allows for a `data.replace()` function, but
it can only take basic strings, and only replaces the first occurrence,
leading us to do things like:

```
{{ data.replace('|', ', ').replace('|', ', ').replace('|', ', ') }}
```

...and hope that there won't be more than 3 pipes.

To improve that, provide a replace filter that can be used like so:

```
{{ data | replace:'\\|':', ' }}
```

I haven't updated the version number on the JS files because it's a
backwards compatible change, and it would require updating the default
template to the new version to opt in (both the default `simple.ftl` but
also the HE package default templates, which is tricky as it's not
properly tied to specific Funnelback versions).