Adds support for includingUrl to select HTML with a CSS selector
SUPPORT-2834

also add support for excluding content with a CSS selector:

* cssSelector: CSS selector to use to select the HTML which should be included. The
selected elemement will be the first to match the selector. The HTML used will include
the element and its attributes. When this is enabled The document may be slightly modified
to be a valid HTML document before the cssSelector is applied this includes wrapping in
`html` tags and `body` tags. This is run before regex modifications and before removeByCssSelector.
* removeByCssSelectors: A list of CSS selectors which match elements which should be removed
from the included HTML. The HTML may be slightly modified to be a valid HTML document this includes
wrapping in `html` tags and `body` tags. This modification will happen even if
`cssSelector` is used. The HTML returned after elements are removed will be the HTML
that is inside of the `body` tag.
* keepBodyAndHeader: When `removeByCssSelectors` is used, the included HTML will be from
the HTML that is within the `body`, which may be automatically added. To instead return
the `header` and `body` tags and their contents this should be set to `true`.