Prevent access denied "javax.management.MBeanTrustPermission" "register"
FUN-12442

It looks like if any security manager is installed then java decides to do a
bunch more checks about what can run and one of these checks prevents metrics
from working.

We added a security manager during the time that groovy scripts run however the
assumption was that the security manager is per thread but it is global this
means that if metrics is called while a groovy script is a running an exception
could be raised. Worse still because it is global this can actually lead to
`GroovyScriptSecurityManager` being left on as the security manager. As we
record the parent manager in the new `GroovyScriptSecurityManager` we end up
with a memory leak.

For now we will remove the security manager as at least right now it is causing
us problems.