Speed up loading curator config RNDSUPPORT-3040.
Each curator action and trigger defined previously had bean annotation processing applied, now we don't do that for all classes which do not require it.

This is especially expensive as for a single query every profile is loaded so every curator trigger and action has this bean processing done on every request.

For vcu this adds about 400ms of time, the saving after making this change is about 400ms per query.

This becomes even more important when we do something per profile during an update e.g. the content auditor summary as what we end up with running a query for every profile which loads the profile config for every profile thus we have a n^2 loop. This MR does not solve the n^2 loop.