Support not returning the CSV header on the all results endpoint.
FUN-13566

The option added is `headerandfooter` and by default is set true, when set false headers and footers for the format will not be returned.

For CSV this means the first row that names fields is not returned.
For JSON this means the `[` and `]` are not returned.
For JSONP this means the same as JSON and the call back is not added.