Add faster index date parsing and reading of the file.
Before: 12000ms to parse and read 450,000 date files  so 1.7ms to do 64
After: 2300ms to parse and read 450,000  so 0.33ms to do 64

I guess now it is clearer as to why I would see it taking 1ms in the query
performance section. The change here makes it better, but perhaps not
fantastic if we wanted to have say 1024 components. At least now in that case
it would be 5.28ms vs 27.2ms.