FUN-15107 removing best bets
Best bets have been deprecated and replaced by curator
best bests since version 15.0

Note that we're making a breaking change by removing
a class and property here. We should not publish this
version of the code as-is, instead we should determine
a strategy to indicate deprecation and a new version
with our version numbering scheme first.